Working register based VM

+++ CLI/src/main.cpp
@@ -85,7 +85,7 @@ int main()
 
 				Eril::FunctionHandle h = vm.GetFunctionHandle(input.c_str());
 				auto handle = h(10, 5423.453, "Test", true, &input, 34.f);
-				std::cout << handle.get<int>() << '\n';
+				std::cout << handle.get<double>() << '\n';
 			}
 			printf("Exiting script stage\n\n");
 		} break;
